+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 2014-12-9
 ---------
  - added <code>"rhumsaa/uuid": "2.*"</code> to composer.json because used by <code>Eye4web\ZfcUser\Pm\Entity\Conversation.</code>
- - added Entity\ConversationTest
+ - added Entity\Conversation Test
+ - added  Entity\ConversationReceiver Test